+++ src/renderer/components/titleBar.vue
@@ -34,6 +34,7 @@
       }
     },
     props: {
+      filename: String,
       pathname: String,
       active: Boolean,
       wordCount: Object
@@ -42,10 +43,6 @@
       paths () {
         const pathnameToken = this.pathname.split('/').filter(i => i)
         return pathnameToken.slice(0, pathnameToken.length - 1).slice(-3)
-      },
-      filename () {
-        const pathnameToken = this.pathname.split('/').filter(i => i)
-        return pathnameToken.pop()
       }
     },
     methods: {